requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
langchain>=0.0.300
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
huggingface-hub>=0.16.4